requirements.txt
psycopg2-binary
python-dotenv
fastapi
uvicorn[standard]
pydantic
openai
